Move RenderPlainText to .Shared
These's no reason this can't be exposed for .NET Core as well - making it so.
